refactor: 클래스 디렉터리 변경 HangmanWord -> Hangman 에서 vo로 이동
