+++ src/main/java/src/vo/HangmanWord.java
@@ -1,4 +1,4 @@
-package src.hangmanGame;
+package src.vo;
 
 public enum HangmanWord {
